[Dialogs] Cleaned system & settings dialogs
