+++ lib/modules/settings/widgets/general/configuration/backup.dart
@@ -12,7 +12,7 @@ class SettingsGeneralConfigurationBackupTile extends StatelessWidget {
 
     Future<void> _backup(BuildContext context) async {
         try {
-            List<dynamic> _values = await LSDialogSystem.showBackupConfigurationPrompt(context);
+            List<dynamic> _values = await LSDialogSettings.backupConfiguration(context);
             if(_values[0]) {
                 String data = Export.export();
                 String encrypted = Encryption.encrypt(_values[1], data);
