+++ lib/modules/settings/widgets/general/profile/enabled_tile.dart
@@ -14,7 +14,7 @@ class SettingsGeneralEnabledProfileTile extends StatelessWidget {
     );
 
     Future<void> _changeProfile(BuildContext context) async {
-        List<dynamic> values = await LSDialogSettings.changeProfile(
+        List<dynamic> values = await LSDialogSettings.enabledProfile(
             context,
             Database.profilesBox.keys.map((x) => x as String).toList()..sort((a,b) => a.toLowerCase().compareTo(b.toLowerCase())),
         );
